requirements.txt
beautifulsoup4==4.9.3
certifi==2021.10.8
ChemSpiPy==2.0.0
click==8.0.3
email-validator==1.1.3
Flask==2.0.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Testing==0.8.1
Flask-WTF==0.14.3
itsdangerous==2.0.1
Jinja2==3.0.2
lxml==4.6.3
pandas==1.4.1
pip==21.2.4
pony==0.7.14
psycopg2-binary==2.9.3
PyJWT==2.1.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
pyyaml==6.0
requests==2.26.0
selenium==3.141.0
setuptools==58.0.4
soupsieve==2.3.1
urllib3==1.26.7
Werkzeug==2.0.3
wheel==0.37.1
WTForms==2.3.3
plotly==5.9.0
